+++ pymc/distributions/__init__.py
@@ -113,6 +113,12 @@ from pymc.distributions.timeseries import (
 )
 from pymc.distributions.truncated import Truncated
 
+# no dispatched stats are being initialized
+# isort: off
+import pymc.distributions._stats
+
+# isort: on
+
 __all__ = [
     "Uniform",
     "Flat",
